Regular commit, just improved stuff :-)

+++ app/src/debug/java/com/together/TestData.kt
@@ -14,7 +14,7 @@ object TestData {
             "Nice fruit to eat.",
             true,
             imageUrl =  "https://firebasestorage.googleapis.com/v0/b/fire-one-58ddc.appspot.com/o/images%2Ftmp1309847000452879291.tmp?alt=media&token=7bfd74a2-30fa-48f8-b4ec-e0b6fc29fc57",
-            price = 2.3.toLong(),
+            price = 2.3,
             unit = "kg"
         ),
         Result.Article(
@@ -25,7 +25,7 @@ object TestData {
             "Leicht säuerlicher Apfel.",
             false,
             imageUrl = "https://firebasestorage.googleapis.com/v0/b/fire-one-58ddc.appspot.com/o/images%2Ftmp3425451583087693428.tmp?alt=media&token=1533803b-a0f4-486e-95ad-39a7abf06df1",
-            price = 2.3.toLong(),
+            price = 2.3,
             unit = "kg"
 
         ),
@@ -37,7 +37,7 @@ object TestData {
             "Festkochende Kartoffel vom Hof Apfeltraum.",
             true,
             imageUrl = "https://firebasestorage.googleapis.com/v0/b/fire-one-58ddc.appspot.com/o/images%2F35e1cd8a-458b-4308-b4d1-2d0bbb8c3369.jpg?alt=media&token=f2a50556-fe36-4bfc-81e4-27ae2324bbd2",
-            price = 2.3.toLong(),
+            price = 2.3,
             unit = "kg"
         ),
         Result.Article(
@@ -48,7 +48,7 @@ object TestData {
             "Aus der Umgebung.",
             true,
             imageUrl = "https://firebasestorage.googleapis.com/v0/b/fire-one-58ddc.appspot.com/o/images%2Ftmp5874223768482947196.tmp?alt=media&token=a3ca5626-0ff4-473f-a289-a8cf090c82f5",
-            price = 2.9.toLong(),
+            price = 2.9,
             unit = "Stück"
         ),
         Result.Article(
@@ -59,7 +59,7 @@ object TestData {
             "Vom Biohof Waldheide",
             true,
             imageUrl = "https://firebasestorage.googleapis.com/v0/b/fire-one-58ddc.appspot.com/o/images%2Ftmp2181782377258883824.tmp?alt=media&token=af827a2e-57ee-4542-9574-25e677c94ba8",
-            price = 3.69.toLong(),
+            price = 3.69,
             unit = "Bund"
         ),
         Result.Article(
@@ -70,7 +70,7 @@ object TestData {
             "Vom Biohof Waldheide",
             true,
             imageUrl = "https://firebasestorage.googleapis.com/v0/b/fire-one-58ddc.appspot.com/o/images%2Ftmp1145260240680560593.tmp?alt=media&token=0670c0da-e260-4d41-b5d8-1a119ea24a64",
-            price = 5.69.toLong(),
+            price = 5.69,
             unit = "Schale"
         ),
         Result.Article(
@@ -81,7 +81,7 @@ object TestData {
             "Vom Biohof Waldheide",
             true,
             imageUrl = "https://firebasestorage.googleapis.com/v0/b/fire-one-58ddc.appspot.com/o/images%2Ftmp6508924231056424774.tmp?alt=media&token=00aae628-6dea-404c-bd7d-101de3ecdd4e",
-            price = 10.69.toLong(),
+            price = 10.69,
             unit = "kg"
         )
     )
